Add files via upload
This code was altered due to containing sensitive information. This codes changes the data into a format that we desire. Sends out emails based on a masterlist from a sharepoint excel file. It has a possibility of having as many emails/customers you want to send out emails to instantly.